requirements.txt
transformers
pandas
torch
numpy
datasets
scikit_learn
notebook
nltk
sklearn-hierarchical-classification